Fix GRC YML blocks
I have spotted two problems with the GRC YML blocks in this PR. First,
with GNU Radio 3.9.2.0 I would get the following errors when generating
Python code with GRC:

Generate Error: (TypeError("'>' not supported between instances of
'TemplateArg' and 'int'")

The fix for this is to use context.get('...')() as for instance in
the UHD blocks.

The second problem is that indentation in the 'make' part of the
template will be translated as is to the generated Python code,
producing code with the wrong indentation. Apparently it is not
possible to indent the lines in the 'make' section, even though that
makes the % if's less readable.

Finally there is a '% end if' -> '% endif' that was already spotted by
@dennisog.